Separate the View part from the EditUserProfileUI comp
